% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_undefined_variables.R
\name{add_undefined_vars}
\alias{add_undefined_vars}
\title{Add missing (undefined) variables to the data dictionary.}
\usage{
add_undefined_vars(dat_dict, undefined_vars = NULL, undefined_note = NULL)
}
\arguments{
\item{dat_dict}{data dictionary as read in with \code{dd_readr()}.}

\item{undefined_vars}{The list of variables to add (which presumably aren't
in the data dictionary.}

\item{undefined_note}{The note to add for undefined variables.  Defaults to
"Not defined - this row was added to the data dictionary as a processing
step".}
}
\value{
\code{dat_dict} with the undefined variables added in.
}
\description{
Add missing (undefined) variables to the data dictionary.
}
