% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dv_tab_cpt.R
\name{build_dv_tab_cpt}
\alias{build_dv_tab_cpt}
\title{Build derived variable table for cancer panel test form}
\usage{
build_dv_tab_cpt(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Cancer panel test form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to cpt with additional derivations.
}
\description{
Build derived variable table for cancer panel test form
}
\examples{
# build_dv_tab_cpt(cpt, dat_dict_sub)
}
