% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dv_tab_med_onc.R
\name{build_dv_tab_med_onc}
\alias{build_dv_tab_med_onc}
\title{Build derived variable table for medical oncologist form}
\usage{
build_dv_tab_med_onc(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Medical oncologist form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to med_onc with additional derivations.
}
\description{
Build derived variable table for medical oncologist form
}
\examples{
# build_dv_tab_med_onc(med_onc, dat_dict_sub)
}
