% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dv_tab_path.R
\name{build_dv_tab_path}
\alias{build_dv_tab_path}
\title{Build derived variable table for pathology form}
\usage{
build_dv_tab_path(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Pathology form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to path with additional derivations.
}
\description{
Build derived variable table for pathology form
}
\examples{
# build_dv_tab_path(path, dat_dict_sub)
}
