% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dv_tab_pt.R
\name{build_dv_tab_pt}
\alias{build_dv_tab_pt}
\title{Build derived variable table for patient form}
\usage{
build_dv_tab_pt(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Curated pt table.}

\item{dat_dict_sub}{Data dictionary with columns relevant to pt table.}
}
\value{
Derived patient table.
}
\description{
Build derived variable table for patient form
}
