% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dv_tab_rad.R
\name{build_dv_tab_rad}
\alias{build_dv_tab_rad}
\title{Build derived variable table for radiation therapy form}
\usage{
build_dv_tab_rad(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Radiation therapy form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to rad with additional derivations.
}
\description{
Build derived variable table for radiation therapy form
}
\examples{
# build_dv_tab_rad(rad, dat_dict_sub)
}
