% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dv_tab_tum.R
\name{build_dv_tab_tum}
\alias{build_dv_tab_tum}
\title{Build derived variable table for tumor marker form}
\usage{
build_dv_tab_tum(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Tumor marker form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to tum with additional derivations.
}
\description{
Build derived variable table for tumor marker form
}
\examples{
# build_dv_tab_tum(tum, dat_dict_sub)
}
