% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_data_derivation_operations.R
\name{common_data_derivation_operations}
\alias{common_data_derivation_operations}
\title{Common data derivation operations}
\usage{
common_data_derivation_operations(
  dat,
  dict,
  print_cols = F,
  exclude_cols = NULL
)
}
\arguments{
\item{dat}{A table representing one PRISSMM instrument}

\item{dict}{The aligned data dictionary for this project.}

\item{print_cols}{Passed on to
\code{convert_dat_col_types()}.}

\item{exclude_cols}{Columns to exclude from the data dictionary before doing anything.}
}
\value{
A dataset with the operations listed above completed.
}
\description{
Currently does two things:  (1) Converts numeric columns to the value stored in the data dictionary. (2) Casts columns that are listed in as numeric in the data dictionary.
}
