% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dat_col_types.R
\name{convert_dat_col_types}
\alias{convert_dat_col_types}
\title{Convert dataset column types}
\usage{
convert_dat_col_types(dat, dict, print_cols = F)
}
\arguments{
\item{dat}{A PRISSMM/REDcap derived dataset.}

\item{dict}{The data dictionary for the project.}

\item{print_cols}{T/F - prints a message about the columns.  Defaults to FALSE.}
}
\value{
A version of \code{dat} with the column types updated.
}
\description{
Convert dataset column types
}
