% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dat_num2val.R
\name{convert_dat_num2val}
\alias{convert_dat_num2val}
\title{Convert the numeric values in the data to the values listed in the data dictionary}
\usage{
convert_dat_num2val(
  dat,
  dict,
  col_to_convert = NULL,
  print_cols = T,
  remove_names = F
)
}
\arguments{
\item{dat}{A PRISSMM/REDCap derived dataset.}

\item{dict}{Aligned data dictionary.}

\item{col_to_convert}{The column to convert.}

\item{print_cols}{Print the columns being worked on?  Default to TRUE.}

\item{remove_names}{Passed on to \code{redcap_numbers_to_meaning()}.}
}
\value{
A dataset with the col_to_convert updated according to the dictionary values.
}
\description{
Convert the numeric values in the data to the values listed in the data dictionary
}
