% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mos_yrs_intervals.R
\name{days_to_yrs}
\alias{days_to_yrs}
\title{Convert days to years}
\usage{
days_to_yrs(x, days_per_yr = 365.25)
}
\arguments{
\item{x}{Numeric vector of days.}

\item{days_per_yr}{Conversion constant. Default 365.25 (from OLD_CODE).}
}
\value{
Numeric vector of years.
}
\description{
Convert days to years
}
