% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_assign_coltypes.R
\name{dd_assign_coltypes}
\alias{dd_assign_coltypes}
\title{Assign coltypes to the data dictionary}
\usage{
dd_assign_coltypes(dd, dttm_cols = NULL, date_cols = NULL, num_cols = NULL)
}
\arguments{
\item{dd}{Data dictionary as read by \code{dd_readr()}.}

\item{dttm_cols}{Columns to cast as datetime (default is along list mostly based ending in _stop_time or _start_time).}

\item{date_cols}{Columns to cast as dates (default is cpt_seq_date and qa_full_date).}

\item{num_cols}{Columns to case as numeric (default is a long list)}
}
\value{
A verison of \code{dd} with \code{col_read_type} column added.
}
\description{
Assign coltypes to the data dictionary
}
