% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_stub_variables.R
\name{expand_stub_variables}
\alias{expand_stub_variables}
\title{Expand stub variables}
\usage{
expand_stub_variables(dd, dat_cols)
}
\arguments{
\item{dd}{A PRISSMM REDCap data dictionary loaded with \code{dd_readr()}.}

\item{dat_cols}{The data columns.}
}
\value{
A tibble with a suitable format to be bound to \code{dd}.
}
\description{
Find stub variables (those following the stub___\link[pillar:num]{pillar::num} format)
and expand them into rows to be added to the data dictionary.
}
