% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_str.R
\name{find_closest_str}
\alias{find_closest_str}
\title{Find the closest string match}
\usage{
find_closest_str(str, dict, method = "lcs")
}
\arguments{
\item{str}{string(s) to match}

\item{dict}{the selection of strings to match to.}

\item{method}{passed on to \code{stringdist::amatch()}}
}
\value{
The closest match for each \code{str}
}
\description{
Find the closest string match
}
