% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{:=}
\alias{:=}
\title{Injection operator}
\usage{
x := y
}
\arguments{
\item{x}{Name of the new variable.}

\item{y}{An expression of what value to assign to the new variable.}
}
\value{
Similar to a = b except it works in special cases.
}
\description{
See \code{rlang::\link[rlang:inject]{:=}} for details.
}
\keyword{internal}
