% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_clin_group_stage.R
\name{map_ca_clin_group_stage}
\alias{map_ca_clin_group_stage}
\title{Map ca_clin_group_stage codes to labels}
\usage{
map_ca_clin_group_stage(x)
}
\arguments{
\item{x}{A vector of ca_clin_group_stage codes.}
}
\value{
A character vector of mapped labels.
}
\description{
Map ca_clin_group_stage codes to labels
}
\examples{
# map_ca_clin_group_stage(ca_dx$ca_clin_group_stage)
}
