% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_clin_n_stage.R
\name{map_ca_clin_n_stage}
\alias{map_ca_clin_n_stage}
\title{Map ca_clin_n_stage codes to text labels}
\usage{
map_ca_clin_n_stage(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{ca_clin_n_stage} codes.}
}
\value{
A character vector of labels.
}
\description{
Map ca_clin_n_stage codes to text labels
}
\examples{
map_ca_clin_n_stage(c(1, 2, 5, 9, 13, 17, 18, 98, 99))
}
