% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_clin_t4_det.R
\name{map_ca_clin_t4_det}
\alias{map_ca_clin_t4_det}
\title{Map ca_clin_t4_det codes to text labels}
\usage{
map_ca_clin_t4_det(x)
}
\arguments{
\item{x}{A numeric (or character) vector of ca_clin_t4_det codes.}
}
\value{
A character vector of clinical T4 detail labels.
}
\description{
Map ca_clin_t4_det codes to text labels
}
\examples{
map_ca_clin_t4_det(c(30, 33, 99))
}
