% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_d_site.R
\name{map_ca_d_site}
\alias{map_ca_d_site}
\title{Map ca_d_site codes to ICD-O-3 text labels}
\usage{
map_ca_d_site(x)
}
\arguments{
\item{x}{A numeric (or character) vector of ca_d_site codes.}
}
\value{
A character vector of ICD-O-3 site labels.
}
\description{
Map ca_d_site codes to ICD-O-3 text labels
}
\examples{
map_ca_d_site(c(340, 500, 809))
}
