% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_dmets_yn.R
\name{map_ca_dmets_yn}
\alias{map_ca_dmets_yn}
\title{Map ca_dmets_yn codes to text labels}
\usage{
map_ca_dmets_yn(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{ca_dmets_yn} codes.}
}
\value{
A character vector of labels.
}
\description{
Map ca_dmets_yn codes to text labels
}
\examples{
map_ca_dmets_yn(c(1, 0, 10))
}
