% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_dx_how.R
\name{map_ca_dx_how}
\alias{map_ca_dx_how}
\title{Map ca_dx_how codes to text labels}
\usage{
map_ca_dx_how(x)
}
\arguments{
\item{x}{A numeric (or character) vector of ca_dx_how codes.}
}
\value{
A character vector of diagnosis method labels.
}
\description{
Map ca_dx_how codes to text labels
}
\examples{
map_ca_dx_how(c(1, 2, 3, 4))
}
