% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_lung_cigarette.R
\name{map_ca_lung_cigarette}
\alias{map_ca_lung_cigarette}
\title{Map ca_lung_cigarette codes to text labels}
\usage{
map_ca_lung_cigarette(x)
}
\arguments{
\item{x}{A numeric (or character) vector of ca_lung_cigarette codes.}
}
\value{
A character vector of cigarette use status labels.
}
\description{
Map ca_lung_cigarette codes to text labels
}
\examples{
map_ca_lung_cigarette(c(1, 2, 3, 4, 5, 100))
}
