% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_path_group_stage.R
\name{map_ca_path_group_stage}
\alias{map_ca_path_group_stage}
\title{Map ca_path_group_stage codes to text labels}
\usage{
map_ca_path_group_stage(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{ca_path_group_stage} codes.}
}
\value{
A character vector of labels.
}
\description{
Map ca_path_group_stage codes to text labels
}
\examples{
map_ca_path_group_stage(c(0, 3, 24, 98, 99))
}
