% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_tx_pre_path_stage.R
\name{map_ca_tx_pre_path_stage}
\alias{map_ca_tx_pre_path_stage}
\title{Map ca_tx_pre_path_stage codes to text labels}
\usage{
map_ca_tx_pre_path_stage(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{ca_tx_pre_path_stage} codes.}
}
\value{
A character vector of labels.
}
\description{
Map ca_tx_pre_path_stage codes to text labels
}
\examples{
map_ca_tx_pre_path_stage(c(1, 0, 98, 99))
}
