% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_type.R
\name{map_ca_type}
\alias{map_ca_type}
\title{Map ca_type codes to text labels}
\usage{
map_ca_type(x)
}
\arguments{
\item{x}{A numeric (or character) vector of ca_type codes.}
}
\value{
A character vector of cancer type labels.
}
\description{
Map ca_type codes to text labels
}
\examples{
map_ca_type(c(1, 13, 41, 1000))
}
