% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_drugs_dc_ynu.R
\name{map_drugs_dc_ynu}
\alias{map_drugs_dc_ynu}
\title{Map drugs_dc_ynu codes to text labels}
\usage{
map_drugs_dc_ynu(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{drugs_dc_ynu} codes.}
}
\value{
A character vector of labels.
}
\description{
Map drugs_dc_ynu codes to text labels
}
\examples{
map_drugs_dc_ynu(c(1, 0, 10))
}
