% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_drugs_firstinst.R
\name{map_drugs_firstinst}
\alias{map_drugs_firstinst}
\title{Map drugs_firstinst codes to text labels}
\usage{
map_drugs_firstinst(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{drugs_firstinst} codes.}
}
\value{
A character vector of labels.
}
\description{
Map drugs_firstinst codes to text labels
}
\examples{
map_drugs_firstinst(c(1, 2))
}
