% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_drugs_inst.R
\name{map_drugs_inst}
\alias{map_drugs_inst}
\title{Map drugs_inst codes to text labels}
\usage{
map_drugs_inst(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{drugs_inst} codes.}
}
\value{
A character vector of labels.
}
\description{
Map drugs_inst codes to text labels
}
\examples{
map_drugs_inst(c(1, 2, 3))
}
