% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_hybrid_death_source.R
\name{map_hybrid_death_source}
\alias{map_hybrid_death_source}
\title{Map hybrid death source codes to text labels}
\usage{
map_hybrid_death_source(x)
}
\arguments{
\item{x}{A numeric (or character) vector of hybrid death source codes.}
}
\value{
A character vector of death source labels.
}
\description{
Map hybrid death source codes to text labels
}
\examples{
map_hybrid_death_source(c(1, 2, 3, 4, 5))
}
