% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_image_inst.R
\name{map_image_inst}
\alias{map_image_inst}
\title{Map image_inst codes to text labels}
\usage{
map_image_inst(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{image_inst} codes.}
}
\value{
A character vector of labels.
}
\description{
Map image_inst codes to text labels
}
\examples{
map_image_inst(c(1, 2))
}
