% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_image_overall.R
\name{map_image_overall}
\alias{map_image_overall}
\title{Map image_overall codes to text labels}
\usage{
map_image_overall(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{image_overall} codes.}
}
\value{
A character vector of labels.
}
\description{
Map image_overall codes to text labels
}
\examples{
map_image_overall(c(1, 2, 3, 4, 5))
}
