% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_naaccr_eth_code.R
\name{map_naaccr_eth_code}
\alias{map_naaccr_eth_code}
\title{Map NAACCR ethnicity codes to text labels}
\usage{
map_naaccr_eth_code(x)
}
\arguments{
\item{x}{A numeric (or character) vector of NAACCR ethnicity codes.}
}
\value{
A character vector of ethnicity labels.
}
\description{
Map NAACCR ethnicity codes to text labels
}
\examples{
map_naaccr_eth_code(c(0, 1, 8, 9))
}
