% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_naaccr_race.R
\name{map_naaccr_race}
\alias{map_naaccr_race}
\title{Map NAACCR race codes to text labels}
\usage{
map_naaccr_race(x)
}
\arguments{
\item{x}{A numeric (or character) vector of NAACCR race codes.}
}
\value{
A character vector of race labels.
}
\description{
Map NAACCR race codes to text labels
}
\examples{
map_naaccr_race(c(1, 2, 4, 99))
}
