% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_path_ca_type.R
\name{map_path_ca_type}
\alias{map_path_ca_type}
\title{Map path_ca_type codes to text labels}
\usage{
map_path_ca_type(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{path_ca_type} codes.}
}
\value{
A character vector of labels.
}
\description{
Map path_ca_type codes to text labels
}
\examples{
map_path_ca_type(c(51, 41, 13))
}
