% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_path_insitu.R
\name{map_path_insitu}
\alias{map_path_insitu}
\title{Map path_insitu codes to text labels}
\usage{
map_path_insitu(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{path_insitu} codes.}
}
\value{
A character vector of labels.
}
\description{
Map path_insitu codes to text labels
}
\examples{
map_path_insitu(c(0, 1, 10))
}
