% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_path_proc.R
\name{map_path_proc}
\alias{map_path_proc}
\title{Map path_proc codes to text labels}
\usage{
map_path_proc(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{path_proc} codes.}
}
\value{
A character vector of labels.
}
\description{
Map path_proc codes to text labels
}
\examples{
map_path_proc(c(1, 2, 3, 9))
}
