% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_path_proc_margins.R
\name{map_path_proc_margins}
\alias{map_path_proc_margins}
\title{Map path_proc_margins codes to text labels}
\usage{
map_path_proc_margins(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{path_proc_margins} codes.}
}
\value{
A character vector of labels.
}
\description{
Map path_proc_margins codes to text labels
}
\examples{
map_path_proc_margins(c(0, 1, 10))
}
