% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_path_proc_type.R
\name{map_path_proc_type}
\alias{map_path_proc_type}
\title{Map path_proc_type codes to text labels}
\usage{
map_path_proc_type(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{path_proc_type} codes.}
}
\value{
A character vector of labels.
}
\description{
Map path_proc_type codes to text labels
}
\examples{
map_path_proc_type(c(1, 2, 3))
}
