% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_path_site.R
\name{map_path_site}
\alias{map_path_site}
\title{Map path_site codes to ICD-O-3 site labels}
\usage{
map_path_site(x)
}
\arguments{
\item{x}{A numeric (or character) vector of site codes.}
}
\value{
A character vector of labels.
}
\description{
Uses the metastasis site mapping from \code{\link[=load_met_site_map]{load_met_site_map()}} to recode
\code{path_site} variables.
}
\examples{
map_path_site(c(349, 809))
}
