% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_valid_value_sets.R
\name{parse_valid_value_sets}
\alias{parse_valid_value_sets}
\title{Parse the PRISSMM data dictionary valid value sets.}
\usage{
parse_valid_value_sets(str, add_na = T)
}
\arguments{
\item{str}{valid value strings (key/val sep by "," and keys separated by |)}

\item{add_na}{Add NA to the valid value sets?}
}
\value{
A named list where the keys are the names and the values are the list values.
}
\description{
Takes the valid value format in prissmm and turns it into named
(magic number keys) character vectors (text values).
}
