% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate_ca_seq.R
\name{propagate_ca_seq}
\alias{propagate_ca_seq}
\title{Propagate the cancer sequence derivation from the cancer dx tables to the others}
\usage{
propagate_ca_seq(lst)
}
\arguments{
\item{lst}{A list of dataframes representing PRISSMM tables. The cancer index table must be called \code{ca_ind} and the non-index cancer table must be called \code{ca_non_ind}. Other than that it's based on column presence/absence.}
}
\value{
\code{lst} with the \code{ca_seq} column added to relevant dataframes.
}
\description{
Propagate the cancer sequence derivation from the cancer dx tables to the others
}
