% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_valid_values.R
\name{split_valid_values}
\alias{split_valid_values}
\title{Split the valid values into structured columns.}
\usage{
split_valid_values(data_dictionary, add_to_existing = T)
}
\arguments{
\item{data_dictionary}{The data dictionary with \code{valid_val_str} already added (for exmaple with \code{dd_readr()}).}

\item{add_to_existing}{Add the derived columns to the existing data?}
}
\value{
A tibble like \code{data_dictionary} with the new columns if add_to_existing = TRUE.  Otherwise just the new cols.
}
\description{
Split the valid values into structured columns.
}
