% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_nameless_cols.R
\name{trim_nameless_cols}
\alias{trim_nameless_cols}
\title{Trim columns with auto-gen names (starting with ...)}
\usage{
trim_nameless_cols(dat)
}
\arguments{
\item{dat}{A loaded dataset.}
}
\value{
\code{dat} with columns starting with "..." removed.
}
\description{
Trim columns with auto-gen names (starting with ...)
}
